requirements.txt
Flask
slack_bolt
requests
python-dotenv